pyproject.toml
[tool.poetry]
name = "libpycom"
version = "0.1.0"
description = ""
authors = ["Kurayuri <Kurayuri@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0"
numpy = "*"
websocket-client = "*"
pyyaml = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
